Prevent baking environment variables into the JAR during compilation
If `metabase.config/config-str` is called during compilation, the value of
an environment variable is baked into the JAR, with potentially harmful
consequences (cf. #30518).  By using `defn`, `(delay ,,,)` or
`(when-not *compile-files* ,,,)` we prevent that.  To prevent accidental
uses in the future we throw if the function is called at compile time.

Initializing the database connection or cache backend during compilation
seemed wrong in any case, since we do not have a proper environment for them
to work, so we make sure to skip that at compile time.

Replaces: https://github.com/metabase/metabase/pull/31612
Closes: https://github.com/metabase/metabase/issues/30518